[feat] FT-154 chat interface에서 관전 기능 개발
