+++ srcs/front/src/components/game/Game.tsx
@@ -33,6 +33,7 @@ function Game() {
 		if (isLoading || !userData || error) return () => {};
 		socket = io('http://3.39.20.24:3032/api/games');
 		socket = socket.on('connect', () => {
+			console.log('connect here');
 			socket.emit('handleUserConnect', userData);
 			socket.emit('getCurrentGames');
 		});
